Moved detail view section to sections folder
